Limpiando el agregar cita
